requirements.txt
Flask==2.0.2
python-telegram-bot==13.7
requests==2.26.0
beautifulsoup4==4.10.0
Werkzeug==2.0.2
python-dotenv==0.19.1
